Extract callback validation utils to own module
